refactor: package image/collection/deck services with mixins (#426)
Convert `services/image_service.py` (563 LOC, multi-responsibility) into a
package whose `ImageService` class is composed from focused mixins:
`bulk_data`, `metadata`, `printing_index`, and `cache`, plus the
`download_queue` helper class.

Consolidate the prior flat `collection_*` submodules into a proper
`services/collection_service/` package with `CollectionService` inheriting
from `CollectionCacheMixin`, `OwnershipMixin`, `DeckAnalysisMixin`,
`StatsMixin`, `BridgeRefreshMixin`, and `ExporterMixin`. The old flat
helper files (`collection_bridge_refresh.py`, `collection_cache.py`, ...)
had no external consumers and were removed.

Move `deck_service.py`, `deck_parser.py`, `deck_averager.py`, and
`deck_text_builder.py` into `services/deck_service/`. Composition between
`DeckService` / `DeckParser` / `DeckAverager` / `DeckTextBuilder` is
preserved; `services/deck_averager.py` etc. are now thin re-export shims
so `from services.deck_averager import DeckAverager` keeps working for
tests.

External callers (`controllers/app_controller.py`, `services/__init__.py`,
`services/radar_service.py`, `widgets/panels/deck_stats_panel.py`, tests)
import only from the package roots and continue to work unchanged.
